src/requirements/base.txt
Django==1.8.19
Pillow==2.9.0
bleach==1.5.0
django-autoslug==1.9.3
django-bleach==0.3.0
django-floppyforms==1.7.0
django-recaptcha==1.1.0
django-sendfile==0.3.11
django-user-sessions==1.3.1
psycopg2==2.7.5
uWSGI==2.0.17.1
